pyproject.toml
[tool.poetry]
name = "smol_dev"
version = "0.0.3"
description = "python module of smol developer"
authors = ["swyx <swyx@dontemail.me>"]
license = "MIT"
readme = "readme.md"
packages = [{ include = "smol_dev" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
openai = "^0.27.8"
openai-function-call = "^0.0.5"
tenacity = "^8.2.2"
agent-protocol = "^1.0.0"
aiohttp = "^3.8.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
src = "src.__main__:main"
api = "smol_dev.api:main"

[tool.poetry.urls]
"Homepage" = "https://github.com/smol-ai/developer"
"Bug Tracker" = "https://github.com/smol-ai/developer/issues"
